+++ lib/src/io/env_parser.dart
@@ -30,8 +30,11 @@ void loadEnv([String path = defaultConfigEnvPath]) {
   }
   configPath = p.canonicalize(path);
   var doc = loadYaml(data);
-  config.intlEnabled = doc['intl']?['enabled'] ?? false;
-  config.outputJsonDir = doc['output_json_dir'] ?? '';
+  // config.intlEnabled = doc['intl']?['enabled'] ?? false;
+  // config.outputJsonDir = doc['output_json_dir'] ?? '';
+  config.outputJsonTemplate = doc['output_json_template'] ?? '';
+  config.outputArbTemplate = doc['output_arb_template'] ?? '';
+  config._configOutputTemplates();
 
   ///'output/assets/l10n'
   config.entryFile = doc['entry_file'] ?? '';
@@ -192,7 +195,7 @@ void _parseSheets(YamlMap doc) {
   var credentials = doc['credentials'];
   var credentialsPath = doc['credentials_path'];
   setCredentials(path: credentialsPath, json: credentials);
-  if (!config.isValidCredentials()){
+  if (!config.isValidCredentials()) {
     trace(
         '''$defaultConfigEnvPath: Please be sure to include your Google Sheet credentials.
 - Use [gsheets:credentials_path:] for a path to your credentials.
@@ -226,12 +229,19 @@ void setCredentials({String? path, Map? json}) {
 class EnvConfig {
   // String outputDir = 'output';
   String dartOutputDir = '';
+
+  ///@deprecated
   String outputJsonDir = '';
+  String outputJsonTemplate = '';
+  String outputArbTemplate = '';
+
   String entryFile = '';
   String paramOutputPattern = '{*}';
   String paramOutputPattern1 = '{{';
   String paramOutputPattern2 = '}}';
-  bool intlEnabled = false;
+
+  /// @deprecated
+  // bool intlEnabled = false;
 
   /// assigned from [entryFile]
   String inputYamlDir = '';
@@ -243,9 +253,9 @@ class EnvConfig {
     return p.canonicalize(p.join(configProjectDir, 'ios'));
   }
 
-  String get intlYamlPath {
-    return !intlEnabled ? '' : joinDir([configProjectDir, 'l10n.yaml']);
-  }
+  // String get intlYamlPath {
+  //   return !intlEnabled ? '' : joinDir([configProjectDir, 'l10n.yaml']);
+  // }
 
   /// "outputDir + tkeys.dart"
   String dartTKeysId = '';
@@ -293,6 +303,7 @@ class EnvConfig {
   String get inputVarsFile => joinDir([config.inputYamlDir, 'vars.lock']);
 
   bool get hasOutputJsonDir => outputJsonDir.isNotEmpty;
+  bool get hasOutputArbDir => outputArbTemplate.isNotEmpty;
 
   bool isValidSheet() =>
       sheetId != null && tableId != null && sheetCredentials != null;
@@ -316,5 +327,59 @@ Please, set [dart:output_dir:] to use any dart generation capability.
       return true;
     }
     return true;
+  }
+
+  String arbOutputDir = '';
+  String _arbOutputFilename = '';
+
+  String jsonOutputDir = '';
+  String _jsonOutputFilename = '';
+
+  void _configOutputTemplates() {
+    if (outputJsonTemplate.isNotEmpty) {
+      var str = outputJsonTemplate;
+      /// json for filename and dir.
+      jsonOutputDir = p.canonicalize(p.dirname(str));
+      /// todo: override for now, clear later.
+      outputJsonDir = jsonOutputDir;
+      _jsonOutputFilename = p.basename(str);
+      if (_jsonOutputFilename.split('*').length > 2) {
+        error(AppStrings.kInvalidOutputJsonTemplate);
+        exit(3);
+      }
+    }
+
+    if (outputArbTemplate.isNotEmpty) {
+      var str = outputArbTemplate;
+      /// arb for filename and dir.
+      arbOutputDir = p.canonicalize(p.dirname(str));
+      _arbOutputFilename = p.basename(str);
+      if (_arbOutputFilename.split('*').length > 2) {
+        error(AppStrings.kInvalidOutputJsonTemplate);
+        exit(3);
+      }
+    }
+  }
+
+  String getArbFilePath(String locale) {
+    if (hasOutputArbDir) {
+      var name = _arbOutputFilename.replaceAll('*', locale);
+      name = p.basenameWithoutExtension(name);
+      /// force extension...
+      name += '.arb';
+      return p.join(arbOutputDir, name);
+    }
+    return '';
+  }
+
+  String getJsonFilePath(String locale) {
+    if (hasOutputJsonDir) {
+      var name = _jsonOutputFilename.replaceAll('*', locale);
+      if (!name.contains('.')) {
+        name += '.json';
+      }
+      return p.join(jsonOutputDir, name);
+    }
+    return '';
   }
 }