+++ lib/src/io/intl_dart_gen.dart
@@ -1,6 +1,3 @@
-import 'package:dcli/dcli.dart';
-import 'package:yaml/yaml.dart';
-
 import '../../flutter_translation_sheet.dart';
 
 // const _pluralKeysValid = {'one', 'zero', 'other', 'many', 'few', 'two'};
@@ -25,7 +22,9 @@ welcomeUser:
     manager: Hello manager!
     other: Hello visitor.
           ''';
+
 Map<String, dynamic> metaFallbackProperties = {};
+Map<String, dynamic> varsByKeys = {};
 
 /// Creates the `{locale}.arb` files to be used by intl package or similar.
 /// when `config.yaml` has `intl:enabled:true`.
@@ -34,19 +33,7 @@ void buildArb(Map<String, Map<String, String>> map) {
   trace('Building arb files');
   var appName = '';
   // detect if we have a l10n.yaml file.
-  var arbDir = _getArbDir();
-  if (arbDir.isEmpty) {
-    error(
-        '''ERROR: $defaultConfigEnvPath, intl package support is enabled [intl:enabled:true]
-But 'arb-dir:' is not define or l10n.yaml not found in your target project.
-Please make sure your trconfig.yaml sits in the root of your project, and l10n.yaml as well.
-
-Follow the Flutter documentation for intl:
-
-https://flutter.dev/docs/development/accessibility-and-localization/internationalization#adding-your-own-localized-messages
-
-''');
-  }
+  // var arbDir = config.arbOutputDir;
 
   /// look for base locales.
   var sublist = config.locales.where((element) => element.contains('_'));
@@ -67,8 +54,9 @@ https://flutter.dev/docs/development/accessibility-and-localization/internationa
       if (appName.isNotEmpty) 'appName': '$appName',
     };
     var localeMap = map[localeKey]!;
-    var pluralMaps = <String, dynamic>{};
-    var selectorMaps = <String, dynamic>{};
+    final pluralMaps = <String, dynamic>{};
+    final selectorMaps = <String, dynamic>{};
+
     // make arb file.
     for (var k in localeMap.keys) {
       if (k.contains(_kSelectorSearch)) {
@@ -78,7 +66,7 @@ https://flutter.dev/docs/development/accessibility-and-localization/internationa
           searchKey: _kSelectorSearch,
           targetMap: selectorMaps,
           errorOnVar: _errorStringInvalidSelector,
-          metaVarType: 'String',
+          // metaVarType: 'String',
         );
       } else if (k.contains(_kPluralSearch)) {
         _customModifier(
@@ -93,7 +81,8 @@ https://flutter.dev/docs/development/accessibility-and-localization/internationa
         var newKey = k.camelCase;
         var textValue = localeMap[k]!;
         output[newKey] = textValue;
-        _addMetaKey(newKey, textValue, output, metaProperties);
+        _addSimpleVarsMetadata(newKey, textValue, output);
+        // _addMetaKey(newKey, textValue, output, metaFallbackProperties);
       }
     }
 
@@ -101,7 +90,6 @@ https://flutter.dev/docs/development/accessibility-and-localization/internationa
     for (var k in pluralMaps.keys) {
       /// add meta keys
       output[k] = _resolvePluralTextFromMap(pluralMaps[k]);
-      trace('a1 ',output[k]);
       _addMetaKey(k, output[k], output, metaProperties);
       _addMetaKey(k, output[k], output, metaFallbackProperties);
     }
@@ -109,18 +97,39 @@ https://flutter.dev/docs/development/accessibility-and-localization/internationa
     /// add selector keys
     for (var k in selectorMaps.keys) {
       output[k] = _resolveSelectorTextFromMap(selectorMaps[k]);
-      trace('a2 ',output[k]);
       _addMetaKey(k, output[k], output, metaProperties);
       _addMetaKey(k, output[k], output, metaFallbackProperties);
     }
-
     var jsonString = prettyJson(output);
-    var outputFilename = 'app_' + localeKey + '.arb';
-    var outputPath = joinDir([arbDir, outputFilename]);
+    var outputPath = config.getArbFilePath(localeKey);
     saveString(outputPath, jsonString);
   }
   trace('arb files generated');
   // runPubGet();
+}
+
+/// generate the metadata key directly from the message if it contains
+/// placeholders.
+void _addSimpleVarsMetadata(
+  String targetKey,
+  String value,
+  Map<String, dynamic> output,
+) {
+  var metaKey = '@' + targetKey;
+  if (varsByKeys.containsKey(targetKey)) {
+    output[metaKey] = <String, dynamic>{
+      'description': 'Auto-generated for $targetKey',
+    };
+
+    /// add the placeholders in the meta key.
+    output[metaKey]['placeholders'] ??= <String, dynamic>{};
+
+    /// clean the placeholder in the original `key`.
+    output[targetKey] = _saveVarsFromString(
+      value,
+      output[metaKey]['placeholders'],
+    );
+  }
 }
 
 /// Parses and adds the needed @metadata for arb files.
@@ -132,8 +141,6 @@ void _customModifier({
   required Map targetMap,
   required String errorOnVar,
 }) {
-  // _addSelector(key: k, value: localeMap[k], targetMap: selectorMaps, errorOnVar: _errorStringInvalidSelector);
-  // var idx1 = key.indexOf(_kSelectorSearch);
   var idx1 = key.indexOf(searchKey);
   var targetKey = key.substring(0, idx1);
   targetKey = targetKey.camelCase;
@@ -145,9 +152,14 @@ void _customModifier({
     return;
   }
   if (!(targetMap[targetKey] as Map).containsKey('var')) {
-    targetMap[targetKey]['var'] = mainToken.split(':').last;
+    final _parts = mainToken.split(':');
+    targetMap[targetKey]['var'] = _parts[1];
+    if (_parts.length > 2) {
+      targetMap[targetKey]['varType'] = _parts[2];
+    }
   }
   String varToken = targetMap[targetKey]['var']!;
+  var varTokenType = targetMap[targetKey]['varType'];
   var selectorKey = keys.last;
   var metaKey = '@' + targetKey;
   metaFallbackProperties[metaKey] ??= <String, dynamic>{
@@ -161,6 +173,10 @@ void _customModifier({
     _map[varToken] = <String, dynamic>{};
     if (metaVarType != null) {
       _map[varToken]!['type'] = metaVarType;
+    } else if (varTokenType != null) {
+      /// fallback type in key definition.
+      /// "selector:gender:String:"
+      _map[varToken]!['type'] = varTokenType;
     }
   }
   var _map = metaFallbackProperties[metaKey]!['placeholders'];
@@ -168,15 +184,25 @@ void _customModifier({
   targetMap[targetKey][selectorKey] = value;
 }
 
+// "sampler": "{timer, select, morning {Order your breakfast now} noon {Order your lunch now} night { Order your dinner now} other {Order your snack now.}}",
+// "@sampler": {
+// "description": "Auto-generated for sampler",
+// "placeholders": {
+// "timer": {}
+// }
+// }
+// }
+
+
 /// Adds a new metakey to [output]
 void _addMetaKey(String newKey, String textValue, Map output, Map metaMap) {
   /// add description always.
   late Map placeholders;
+
   /// check if the property had metadata.
   final metaKey = '@$newKey';
   if (metaMap.containsKey(metaKey)) {
     output[metaKey] = Map.from(metaMap[metaKey]);
-    // trace("Meta map has it! ${metaKey} /// ${metaMap[metaKey]}");
   } else {
     output[metaKey] = <String, dynamic>{};
   }
@@ -188,14 +214,8 @@ void _addMetaKey(String newKey, String textValue, Map output, Map metaMap) {
         metaMap[metaKey].containsKey('placeholders')) {
       placeholders.addAll(metaMap[metaKey]['placeholders']);
     }
-
   }
   _saveVarsFromString(textValue, placeholders);
-  trace("Metmap: ", metaMap);
-  trace("NEW KEY: ", newKey);
-  trace("TEXT: ", textValue);
-  trace("OUTPUT: ", output);
-
 }
 
 const _emptyVarMap = <String, dynamic>{};
@@ -204,17 +224,21 @@ const _emptyVarMap = <String, dynamic>{};
 /// basedir.
 String _saveVarsFromString(String value, Map saveTo) {
   final textVars = _varsFromString(value);
+  // trace("So the::: ", textVars,  ' -- ', value, '-- saveInd:', saveTo);
   var cleanedValue = value;
   if (textVars.isNotEmpty) {
+    // trace("CLeaned pre:: ", cleanedValue, '===', textVars);
     textVars.forEach((key, value) {
       /// take complex {vars:type:format(params)} and replace the value with the simple var name.
       if (value is Map && value.containsKey('_text')) {
         cleanedValue = cleanedValue.replaceAll(value['_text'], key);
         value.remove('_text');
       }
-      return saveTo.putIfAbsent(key, () => value);
+      saveTo[key] = value;
+      // return saveTo.putIfAbsent(key, () => value);
     });
   }
+  // trace("Cleaned value::", cleanedValue);
   return cleanedValue;
 }
 
@@ -225,7 +249,7 @@ Map<String, dynamic> _varsFromString(String text) {
     if (res.isNotEmpty) {
       var output = <String, dynamic>{};
       res.forEach((e) {
-        /// todo: analyze more data in the var type.
+        /// todo: Analyze more data in the var type.
         if (e.contains(':')) {
           var propData = _buildMetaVarProperties(e);
           output[propData['name']] = propData['data'];
@@ -239,6 +263,13 @@ Map<String, dynamic> _varsFromString(String text) {
   return _emptyVarMap;
 }
 
+const _kDefaultVarTypeFormats = <String, String>{
+  'num': 'decimalPattern',
+  'double': 'decimalPattern',
+  'int': 'decimalPattern',
+  'DateTime': 'yMd',
+};
+
 /// capture format/type/parameters from a composed variable {{name:String}}.
 Map<String, dynamic> _buildMetaVarProperties(String text) {
   final parts = text.split(':');
@@ -247,6 +278,11 @@ Map<String, dynamic> _buildMetaVarProperties(String text) {
   var vo = <String, dynamic>{
     'type': type,
   };
+
+  /// analyze if `type` requires a Format for intl generation.
+  if (parts.isEmpty && _kDefaultVarTypeFormats.containsKey(type)) {
+    parts.add(_kDefaultVarTypeFormats[type]!);
+  }
   if (parts.isNotEmpty) {
     var msg = parts.join(':');
     var startIndex = msg.indexOf('(');
@@ -305,10 +341,16 @@ String _resolveSelectorTextFromMap(Map map) {
     error(
         'select: keys must contain the type "$_selectorMandatory" as default.');
   }
+
+  /// Remove `varType` if we have type info inside the field name.
+  /// "selector:gender:String"
+  map.remove('varType');
   var varKey = map.remove('var');
-  str += '{$varKey, select, ';
+  /// Format is critical to be accepted by intl tools.
+  /// NO SPACES at the end.
+  str += '{$varKey, select,';
   for (var k in map.keys) {
-    str += '$k {${map[k]}} ';
+    str += ' $k{${map[k]}}';
   }
   str += '}';
   return str;
@@ -333,19 +375,19 @@ String _resolvePluralTextFromMap(Map map) {
 }
 
 /// Returns the folder to save the `.arb` output.
-String _getArbDir() {
-  var intlPath = config.intlYamlPath;
-  if (intlPath.isNotEmpty && exists(intlPath)) {
-    var data = openString(intlPath);
-    if (data.isEmpty) {
-      return '';
-    }
-    var doc = loadYaml(data);
-    var arbDir = doc['arb-dir'];
-    return joinDir([configProjectDir, arbDir]);
-  }
-  return '';
-}
+// String _getArbDir() {
+//   var intlPath = config.intlYamlPath;
+//   if (intlPath.isNotEmpty && exists(intlPath)) {
+//     var data = openString(intlPath);
+//     if (data.isEmpty) {
+//       return '';
+//     }
+//     var doc = loadYaml(data);
+//     var arbDir = doc['arb-dir'];
+//     return joinDir([configProjectDir, arbDir]);
+//   }
+//   return '';
+// }
 
 final _matchArbPlaceholderRegExp = RegExp(r'\{(.+?)\}');