+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## [1.0.14+31]
+- fixed simple placeholder generation for arb.
+- improved plural and selectors modifiers for arb.
+- changed trconfig.yaml structure: now you can provide [output_json_template] and [output_arb_template] using the * as placeholder for locale replacement. ( For example `output_arb_template: lib/l10n/intl_*.arb`)
+- other minor adjustments.
+
 ## [1.0.14+30]
 - fixed `fts --version` for local development.
 - added global catch() for sheet errors, to provide in the future better hints on the errors and possible solutions.